atet_worker: Move Worker class out of acryd-tester
Extract _Worker class out of arcyd-tester into a new file
atet_worker.py. Not adding unit tests explicitly for this class as all
functions are exercised in arcyd-tester, especially those for the happy
path and unit tests would simply duplicate that.

Test plan:
$ ./precommit.sh
