+++ py/ate/README.md
@@ -3,6 +3,8 @@
 Arcyd tester - command-line utility to exercise arcyd.
 * `atet_arcyd_instance.py` -
 Tools to start and manage an Arcyd instance.
+* `atet_worker.py` -
+Utilities to simulate a person working with Phabricator, Git and Arcyd.
 
 -----
 *please note: this file is generated, edits will be lost*
