+++ py/ate/atecmd_arcydtester.py
@@ -20,7 +20,6 @@ from __future__ import print_function
 import argparse
 import contextlib
 import itertools
-import json
 import os
 import shutil
 import stat
@@ -28,14 +27,12 @@ import subprocess
 import tempfile
 
 import phldef_conduit
-import phlgit_push
-import phlgitu_fixture
 import phlsys_fs
 import phlsys_git
 import phlsys_web
-import phlsys_workingdircommand
 
 import atet_arcyd_instance
+import atet_worker
 
 _USAGE_EXAMPLES = """
 """
@@ -122,97 +119,6 @@ def main():
         _do_tests(args)
 
 
-class _Worker(object):
-
-    """Simulate a person working with Phabricator, Git and Arcyd."""
-
-    def __init__(
-            self,
-            repo,
-            working_dir,
-            barc_cmd_path,
-            phab_username,
-            email,
-            conduit_cert,
-            arcyon_cmd_path,
-            phab_uri):
-        self._repo = repo
-        self._barc = phlsys_workingdircommand.CommandWithWorkingDirectory(
-            barc_cmd_path, working_dir)
-        self._arcyon = phlsys_workingdircommand.CommandWithWorkingDirectory(
-            arcyon_cmd_path, working_dir)
-        self._working_dir = working_dir
-        self._git_worker = phlgitu_fixture.Worker(repo)
-        self._phab_username = phab_username
-        self._email = email
-        self._conduit_cert = conduit_cert
-        self._phab_uri = phab_uri
-
-    def setup(self, central_repo_path):
-        self._repo("init")
-        self._repo('config', 'user.name', self._phab_username)
-        self._repo('config', 'user.email', self._email)
-        self._repo("remote", "add", "origin", central_repo_path)
-        self._repo("fetch")
-
-    def push_initial_commit(self):
-        self._git_worker.commit_new_file('initial commit', 'README')
-        phlgit_push.push(self._repo, 'master', 'origin')
-
-    def push_new_review_branch(self, identifier, message=None):
-        branch_name = 'r/master/{}'.format(identifier)
-
-        if message is None:
-            message = 'Making review for {}'.format(identifier)
-
-        self._git_worker.commit_new_file_on_new_branch(
-            branch=branch_name,
-            message=message,
-            relative_path=identifier,
-            base='origin/master')
-        return phlgit_push.push(self._repo, branch_name, 'origin')
-
-    def push_review_update(self, identifier, to_append):
-        self._git_worker.commit_append_to_file(
-            identifier, identifier, to_append)
-
-        branch_name = 'r/master/{}'.format(identifier)
-        phlgit_push.push(self._repo, branch_name, 'origin')
-
-    def fetch(self):
-        self._repo('fetch', '--prune')
-
-    def list_reviews(self):
-        return json.loads(self._barc('list', '--format-json'))
-
-    def _arcyon_action(self, review_id, action):
-        connection_args = [
-            '--user', self._phab_username,
-            '--cert', self._conduit_cert,
-            '--uri', self._phab_uri,
-        ]
-        self._arcyon(
-            'comment',
-            review_id,
-            '--message', 'accepting',
-            '--action', action,
-            *connection_args)
-
-    def accept_review(self, review_id):
-        self._arcyon_action(review_id, 'accept')
-
-    def request_changes(self, review_id):
-        self._arcyon_action(review_id, 'request changes')
-
-    @property
-    def repo(self):
-        return self._repo
-
-    @property
-    def barc(self):
-        return self._barc
-
-
 class _SharedRepo(object):
 
     def __init__(
@@ -256,7 +162,7 @@ class _SharedRepo(object):
             worker_path = os.path.join(self._root_dir, account_user)
             os.makedirs(worker_path)
             self._workers.append(
-                _Worker(
+                atet_worker.Worker(
                     phlsys_git.Repo(worker_path),
                     worker_path,
                     barc_cmd_path,
